backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
python-dotenv==1.0.1
pyodbc==5.1.0
SQLAlchemy==2.0.27
python-dateutil==2.8.2
